TypeTrack Pro/requirements.txt
flask
flask-cors
flask-bcrypt
flask-jwt-extended
postgrest-py
httpx
python-dotenv
